requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
windows-curses==2.3.0
yarg==0.1.9
